fix(chaos): integer-cent money, pt-BR amounts, input validation, multibyte editing
Per author decisions after the chaos run:
- Amount is int64 cents everywhere; SQLite gains amount_cents INTEGER with a
  one-shot backfill from the legacy REAL column (kept in sync for old readers)
- ParseCurrency: last separator is decimal, lone comma is decimal (1,5 = 1.50),
  rejects negatives, NaN/Inf, empty and malformed grouping; FormatCurrency is
  total and renders $1.234,56
- Ledger rejects negative amounts, unknown Kind and duplicate IDs; ChartSeries
  guards n<=0 before allocating
- Table editor: invalid date/amount keeps the cell in edit mode and surfaces the
  error in the status bar; key handler uses msg.Runes so ç/ã/emoji work;
  Backspace is rune-aware
- SQLite: busy_timeout(5000) + WAL so two instances coexist; empty path errors
  instead of panicking; unknown entry_type fails Load loudly
- AddMonths clamps to month end (Jan 31 + 1 = Feb 28)

All 19 chaos findings re-armed and passing; 41 chaos tests, 0 skipped.

+++ internal/ui/components/statusbar.go
@@ -11,7 +11,7 @@ import (
 
 type StatusBar struct {
 	TimeRange    string
-	TotalBalance float64
+	TotalBalance int64
 	Err          error
 	Width        int
 }
@@ -26,7 +26,7 @@ func (s *StatusBar) SetTimeRange(start, end string) {
 	s.TimeRange = fmt.Sprintf("%s - %s", start, end)
 }
 
-func (s *StatusBar) SetBalance(balance float64) {
+func (s *StatusBar) SetBalance(balance int64) {
 	s.TotalBalance = balance
 }
